+++ platform-vivadohls/src/main/java/ch/epfl/vlsc/hls/backend/systemc/Simulator.java
@@ -76,7 +76,7 @@ public interface Simulator {
     }
 
     default LogicVector getSCPortType(PortDecl port) {
-        int bitWidth = backend().typeseval().sizeOfBits(backend().types().declaredPortType(port));
+        int bitWidth = (int) backend().typeseval().sizeOfBits(backend().types().declaredPortType(port));
         return new LogicVector(bitWidth);
     }
